+++ Model/Ui/TokenUiComponentProvider.php
@@ -87,10 +87,9 @@ class TokenUiComponentProvider implements TokenUiComponentProviderInterface
 
         // Check for merchant ID, if the token merchant ID does not match the payment extension merchant ID
         // then do not render the vault method in hand.
-        // @todo: not the best way to decide if a token payment needs to be rendered, refactor it
-        //
+        // Not the best way to decide if a token payment needs to be rendered, need to refactor it
         if (!isset($jsonDetails['merchant_id']) || $this->config->getMerchantId() !== $jsonDetails['merchant_id']) {
-            return $component = $this->componentFactory->create([
+           return $this->componentFactory->create([
                 'config' => [],
                 'name' => null
             ]);
@@ -104,13 +103,11 @@ class TokenUiComponentProvider implements TokenUiComponentProviderInterface
             TokenUiComponentProviderInterface::COMPONENT_PUBLIC_HASH => $paymentToken->getPublicHash(),
         ];
 
-        $component = $this->componentFactory->create(
+        return $this->componentFactory->create(
             [
                 'config' => array_merge($config, $vaultConfig->getVaultConfig()),
                 'name' => $vaultConfig->getVaultComponent()
             ]
         );
-
-        return $component;
     }
 }